+++ signalwire/relay/calling/__init__.py
@@ -29,6 +29,8 @@ class Calling(BaseRelay):
       self._on_fax(notification['params'])
     elif notification['event_type'] == Notification.SEND_DIGITS:
       self._on_send_digits(notification['params'])
+    elif notification['event_type'] == Notification.TAP:
+      self._on_tap(notification['params'])
 
   def new_call(self, *, call_type='phone', from_number, to_number, timeout=None):
     call = Call(calling=self)
@@ -129,3 +131,10 @@ class Calling(BaseRelay):
       trigger(Notification.SEND_DIGITS, params, suffix=params['control_id']) # Notify components listening on SendDigits and control_id
       trigger(call.tag, params, suffix='sendDigits.stateChange')
       trigger(call.tag, params, suffix=f"sendDigits.{params['state']}")
+
+  def _on_tap(self, params):
+    call = self._get_call_by_id(params['call_id'])
+    if call is not None:
+      trigger(Notification.TAP, params, suffix=params['control_id']) # Notify components listening on Tap and control_id
+      trigger(call.tag, params, suffix='tap.stateChange')
+      trigger(call.tag, params, suffix=f"tap.{params['state']}")